requirements.txt
accelerate>=0.27.2
transformers>=4.41.2
datasets>=2.19.0
peft>=0.11.1
torch>=2.2.1
pyyaml>=6.0.1
google-generativeai>=0.7.2
tqdm>=4.66.4
sentencepiece>=0.2.0
